+++ imdbinfo/__init__.py
@@ -23,6 +23,7 @@ import logging
 
 from .services import (
     get_movie,
+    get_awards,
     search_title,
     get_name,
     get_episodes,
@@ -39,6 +40,8 @@ from .services import (
     TitleType,
 )
 from .models import (
+    Award,
+    AwardInfo,
     Quote,
     QuoteLine,
     QuoteCharacter,
@@ -54,6 +57,7 @@ from .exceptions import (
 
 __all__ = [
     "get_movie",
+    "get_awards",
     "search_title",
     "get_name",
     "get_episodes",
@@ -68,6 +72,9 @@ __all__ = [
     "get_media_gallery",
     "get_quotes",
     "TitleType",
+    # award models
+    "Award",
+    "AwardInfo",
     # quote models
     "Quote",
     "QuoteLine",